refactor(java): split json codegen owners
